create utils folder. added hooks to util. added fixed nav
