+++ app/routes/projects/repos.tsx
@@ -5,7 +5,7 @@ import { IconGitFork, IconUserCircle } from "@tabler/icons-react";
 import { getParam, RepoData, sorts, UserScope } from "~/server/projects";
 import { Card, SortControl } from "~/components";
 import commonStyles from "~/styles/common.module.css";
-import { capitalize, validate } from "~/util";
+import { capitalize, validate } from "~/utils";
 
 import Language from "./language";
 import styles from "./repos.module.css";
