+++ app/utils/frontend.ts
@@ -1,0 +1,3 @@
+export function cls(...args: (string | boolean | undefined | null)[]) {
+  return args.filter(Boolean).join(" ").trim();
+}
