+++ app/utils/hooks.ts
@@ -1,0 +1,16 @@
+import { useEffect, useRef, useState } from "react";
+
+export function useElementRect<T extends HTMLElement>() {
+  const ref = useRef<T>(null);
+  const [rect, setRect] = useState<DOMRect | undefined>();
+
+  useEffect(() => {
+    if (ref.current) {
+      setRect(ref.current.getBoundingClientRect());
+    }
+    // add ref.current as dep or else hook won't fire
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [ref.current]);
+
+  return [ref, rect] as const;
+}
