+++ scripts/blog-pull.ts
@@ -3,7 +3,7 @@ import fs from "node:fs/promises";
 import path from "path";
 
 import DB from "~/server/db.singleton.server";
-import { exists, kebobCase } from "~/util";
+import { exists, kebobCase } from "~/utils";
 
 const dir = path.resolve(".", "app/blog");
